Use a mocked Sys::Syslog for main log tests.
This addresses the problem in GH#1.

An explanation of the problem:

The tests assumed that calling any log methods would return whatever the
call to `Sys::Syslog::syslog` returns. This was fine until Dancer2 commit
c56951f8a8a6987aaf4a7fd467fa0c9f403f748c which added engine.logger.*
hooks around logging a message and so the return value of `log` and
thefore also of the called `syslog` became lost causing tests to fail.

Link to commit on GH:
https://github.com/PerlDancer/Dancer2/commit/c56951f8a8a6987aaf4a7fd467fa0c9f403f748c

In order to allow tests to work this commit adds a fake local
implementation of `Sys::Syslog` which simply calls `warn` with whatever
list was passed into `syslog`. We catch and test the warnings with
`Test::Warnings`.